refactor: api 파일 정리 및 image 관련 api 수정
