+++ app/apis/images/index.ts
@@ -1,7 +1,7 @@
 'use server';
 
-import { storageService } from '@/firebase';
-import { getDownloadURL, ref } from 'firebase/storage';
+import { authService, storageService } from '@/firebase';
+import { getDownloadURL, ref, uploadString } from 'firebase/storage';
 
 export const getImageURL = async (url: string) => {
   const photoRef = ref(storageService, `${url}.jpg`);
@@ -24,4 +24,16 @@ export const getPostImageURL = async (
   const downloadURL = await getDownloadURL(photoRef);
 
   return downloadURL;
+};
+
+export const uploadImage = async (path: string, url: string) => {
+  const user = authService.currentUser;
+
+  if (!user || !url) return;
+
+  const photoRef = ref(storageService, `/${path}.jpg`);
+  await uploadString(photoRef, url, 'data_url');
+
+  const photoURL = await getDownloadURL(photoRef);
+  return photoURL;
 };
